style: 调整重复确认逻辑, 优化使用方法，减少不必要的pending状态

+++ web/src/components/ChatBox/ChatMessageList/useChatMessageItemAction.tsx
@@ -25,13 +25,15 @@ export function useChatMessageItemAction(
   const userInfo = useUserInfo();
 
   const [, handleRecallMessage] = useAsyncRequest(async () => {
-    await openReconfirmModalP();
-    await recallMessage(payload._id);
+    if (await openReconfirmModalP()) {
+      await recallMessage(payload._id);
+    }
   }, [payload._id]);
 
   const [, handleDeleteMessage] = useAsyncRequest(async () => {
-    await openReconfirmModalP();
-    await deleteMessage(payload._id);
+    if (await openReconfirmModalP()) {
+      await deleteMessage(payload._id);
+    }
   }, [payload._id]);
 
   const isGroupOwner = groupInfo && groupInfo.owner === userInfo?._id; //
